requirements.txt
arabic-reshaper==2.1.3
asgiref==3.2.7
autopep8==1.6.0
certifi==2020.4.5.1
charset-normalizer==2.0.7
click==8.0.4
colorama==0.4.4
dj-database-url==0.5.0
Django==3.0.7
django-bootstrap3==21.2
django-jquery==3.1.0
django-phone-verify==0.1.1
django-phonenumber-field==6.0.0
django-timezone-field==4.0
djangorestframework==3.11.0
djlint==0.7.5
Faker==4.1.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==3.3
importlib-metadata==4.11.2
pathspec==0.9.0
phonenumbers==8.12.37
Pillow==8.4.0
psycopg2-binary==2.8.6
pycodestyle==2.8.0
PyJWT==2.3.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-dotenv==0.19.2
pytz==2020.1
PyYAML==6.0
regex==2022.3.2
reportlab==3.6.3
requests==2.26.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.2
tomlkit==0.9.2
tqdm==4.63.0
twilio==7.3.1
urllib3==1.26.7
webencodings==0.5.1
whitenoise==5.1.0
wincertstore==0.2
xhtml2pdf==0.2.5
zipp==3.7.0
